# swagger.json
{
    "swagger": "2.0",
    "info": {
      "title": "PORTFOLIO API",
      "description": "A personal portfolio API",
      "version": "1.0.0",
      "contact": {
        "name": "Honore IRADUKUNDA",
        "email": "ihonore01@gmail.com",
        "url": "ihonore.netlify.app" 
      }
    },
    "schemes": ["http", "https"],
    "servers": [
      {
        "url": "http://localhost:3000",
        "name": "DEV"
      }
    ],
    "securityDefinitions": {
      "Bearer": {
        "type": "apiKey",
        "name": "Authorization",
        "in": "header"
      }
    },
    "paths": {
...
      "/api/v1/comments/{articleId}": {
        "get": {
          "tags": ["Comments"],
          "summary": "Get all comments of given article",
          "description": "Get all comments of given article",
          "operationId": "getComments",
          "parameters": [
            {
              "name": "articleId",
              "in": "path",
              "description": " Enter the article ID to get its comments",
              "required": true,
              "type": "string"
            }],
          "responses": {
            "200": {
              "description": "All comments retrieved successfuly!"
            }
          }
        },
        "post": {
          "tags": ["Comments"],
          "summary": "Create a comment",
          "description": "Create a comment",
          "operationId": "createComment",
          "parameters": [
            {
              "name": "articleId",
              "in": "path",
              "description": "Enter the ID of an article your want to comment on",
              "required": true,
              "type": "string"
            },
            {
              "name": "body",
              "in": "body",
              "description": "Comment section",
              "required": true,
              "schema": {
                "$ref": "#/definitions/Comment"
              }
            }
          ],
          "responses": {
            "201": {
              "description": "Comment submitted successfuly!"
            },
            "500": {
              "description": "Server error"
            }
          }
        }
      },
...
    },
    "definitions": {
...
      "Comment": {
        "type": "object",
        "required": ["commenter", "comment"],
        "properties": {
          "commenter": {
            "type": "string"
          },
          "comment": {
            "type": "string"
          }
        }
      }
    }
}